Rework CLI
