+++ cmd/pyroscope/main.go
@@ -1,5 +1,3 @@
-// +build !windows
-
 package main
 
 import (
