+++ pkg/cli/agent_unix.go
@@ -1,0 +1,13 @@
+// +build !windows
+
+package cli
+
+import (
+	"fmt"
+
+	"github.com/pyroscope-io/pyroscope/pkg/config"
+)
+
+func startAgent(_ *config.Agent) error {
+	return fmt.Errorf("agent mode is not supported")
+}
