+++ pkg/cli/cli.go
@@ -4,232 +4,24 @@ import (
 	"context"
 	"flag"
 	"fmt"
+	"io/ioutil"
 	"os"
 	"path/filepath"
-	"reflect"
 	"runtime"
-	"strconv"
-	"strings"
 	"time"
 
-	"github.com/iancoleman/strcase"
 	"github.com/peterbourgon/ff/v3"
 	"github.com/peterbourgon/ff/v3/ffcli"
 	"github.com/sirupsen/logrus"
 	"gopkg.in/yaml.v2"
 
-	"github.com/pyroscope-io/pyroscope/pkg/agent"
-	"github.com/pyroscope-io/pyroscope/pkg/agent/spy"
-	"github.com/pyroscope-io/pyroscope/pkg/agent/upstream/direct"
-	"github.com/pyroscope-io/pyroscope/pkg/analytics"
 	"github.com/pyroscope-io/pyroscope/pkg/build"
 	"github.com/pyroscope-io/pyroscope/pkg/config"
 	"github.com/pyroscope-io/pyroscope/pkg/convert"
 	"github.com/pyroscope-io/pyroscope/pkg/dbmanager"
 	"github.com/pyroscope-io/pyroscope/pkg/exec"
-	"github.com/pyroscope-io/pyroscope/pkg/server"
-	"github.com/pyroscope-io/pyroscope/pkg/storage"
-	"github.com/pyroscope-io/pyroscope/pkg/util/atexit"
-	"github.com/pyroscope-io/pyroscope/pkg/util/bytesize"
 	"github.com/pyroscope-io/pyroscope/pkg/util/debug"
-	"github.com/pyroscope-io/pyroscope/pkg/util/slices"
 )
-
-const timeFormat = "2006-01-02T15:04:05Z0700"
-
-type arrayFlags []string
-
-func (i *arrayFlags) String() string {
-	return strings.Join(*i, ", ")
-}
-
-func (i *arrayFlags) Set(value string) error {
-	*i = append(*i, value)
-	return nil
-}
-
-type timeFlag time.Time
-
-func (tf *timeFlag) String() string {
-	v := time.Time(*tf)
-	return v.Format(timeFormat)
-}
-
-func (tf *timeFlag) Set(value string) error {
-	t2, err := time.Parse(timeFormat, value)
-	if err != nil {
-		var i int
-		i, err = strconv.Atoi(value)
-		if err != nil {
-			return err
-		}
-		t2 = time.Unix(int64(i), 0)
-	}
-
-	t := (*time.Time)(tf)
-	b, _ := t2.MarshalBinary()
-	t.UnmarshalBinary(b)
-
-	return nil
-}
-
-// this is mostly reflection magic
-func PopulateFlagSet(obj interface{}, flagSet *flag.FlagSet, skip ...string) *SortedFlags {
-	v := reflect.ValueOf(obj).Elem()
-	t := reflect.TypeOf(v.Interface())
-	num := t.NumField()
-
-	installPrefix := getInstallPrefix()
-	supportedSpies := strings.Join(spy.SupportedExecSpies(), ", ")
-
-	for i := 0; i < num; i++ {
-		field := t.Field(i)
-		fieldV := v.Field(i)
-		if !(fieldV.IsValid() && fieldV.CanSet()) {
-			continue
-		}
-
-		defaultValStr := field.Tag.Get("def")
-		descVal := field.Tag.Get("desc")
-		skipVal := field.Tag.Get("skip")
-		nameVal := field.Tag.Get("name")
-		if nameVal == "" {
-			nameVal = strcase.ToKebab(field.Name)
-		}
-		if skipVal == "true" || slices.StringContains(skip, nameVal) {
-			continue
-		}
-
-		descVal = strings.ReplaceAll(descVal, "<supportedProfilers>", supportedSpies)
-
-		switch field.Type {
-		case reflect.TypeOf([]string{}):
-			val := fieldV.Addr().Interface().(*[]string)
-			val2 := (*arrayFlags)(val)
-			flagSet.Var(val2, nameVal, descVal)
-		case reflect.TypeOf(""):
-			val := fieldV.Addr().Interface().(*string)
-			defaultValStr := strings.ReplaceAll(defaultValStr, "<installPrefix>", installPrefix)
-			flagSet.StringVar(val, nameVal, defaultValStr, descVal)
-		case reflect.TypeOf(true):
-			val := fieldV.Addr().Interface().(*bool)
-			flagSet.BoolVar(val, nameVal, defaultValStr == "true", descVal)
-		case reflect.TypeOf(time.Time{}):
-			valTime := fieldV.Addr().Interface().(*time.Time)
-			val := (*timeFlag)(valTime)
-			flagSet.Var(val, nameVal, descVal)
-		case reflect.TypeOf(time.Second):
-			val := fieldV.Addr().Interface().(*time.Duration)
-			var defaultVal time.Duration
-			if defaultValStr == "" {
-				defaultVal = time.Duration(0)
-			} else {
-				var err error
-				defaultVal, err = time.ParseDuration(defaultValStr)
-				if err != nil {
-					logrus.Fatalf("invalid default value: %q (%s)", defaultValStr, nameVal)
-				}
-			}
-			flagSet.DurationVar(val, nameVal, defaultVal, descVal)
-		case reflect.TypeOf(bytesize.Byte):
-			val := fieldV.Addr().Interface().(*bytesize.ByteSize)
-			var defaultVal bytesize.ByteSize
-			if defaultValStr != "" {
-				var err error
-				defaultVal, err = bytesize.Parse(defaultValStr)
-				if err != nil {
-					logrus.Fatalf("invalid default value: %q (%s)", defaultValStr, nameVal)
-				}
-			}
-			*val = defaultVal
-			flagSet.Var(val, nameVal, descVal)
-		case reflect.TypeOf(1):
-			val := fieldV.Addr().Interface().(*int)
-			var defaultVal int
-			if defaultValStr == "" {
-				defaultVal = 0
-			} else {
-				var err error
-				defaultVal, err = strconv.Atoi(defaultValStr)
-				if err != nil {
-					logrus.Fatalf("invalid default value: %q (%s)", defaultValStr, nameVal)
-				}
-			}
-			flagSet.IntVar(val, nameVal, defaultVal, descVal)
-		case reflect.TypeOf(1.00):
-			val := fieldV.Addr().Interface().(*float64)
-			var defaultVal float64
-			if defaultValStr == "" {
-				defaultVal = 0.00
-			} else {
-				var err error
-				defaultVal, err = strconv.ParseFloat(defaultValStr, 64)
-				if err != nil {
-					logrus.Fatalf("invalid default value: %q (%s)", defaultValStr, nameVal)
-				}
-			}
-			flagSet.Float64Var(val, nameVal, defaultVal, descVal)
-		case reflect.TypeOf(uint64(1)):
-			val := fieldV.Addr().Interface().(*uint64)
-			var defaultVal uint64
-			if defaultValStr == "" {
-				defaultVal = uint64(0)
-			} else {
-				var err error
-				defaultVal, err = strconv.ParseUint(defaultValStr, 10, 64)
-				if err != nil {
-					logrus.Fatalf("invalid default value: %q (%s)", defaultValStr, nameVal)
-				}
-			}
-			flagSet.Uint64Var(val, nameVal, defaultVal, descVal)
-		case reflect.TypeOf(uint(1)):
-			val := fieldV.Addr().Interface().(*uint)
-			var defaultVal uint
-			if defaultValStr == "" {
-				defaultVal = uint(0)
-			} else {
-				out, err := strconv.ParseUint(defaultValStr, 10, 64)
-				if err != nil {
-					logrus.Fatalf("invalid default value: %q (%s)", defaultValStr, nameVal)
-				}
-				defaultVal = uint(out)
-			}
-			flagSet.UintVar(val, nameVal, defaultVal, descVal)
-		default:
-			logrus.Fatalf("type %s is not supported", field.Type)
-		}
-	}
-	return NewSortedFlags(obj, flagSet)
-}
-
-// on mac pyroscope is usually installed via homebrew. homebrew installs under a prefix
-//   this is logic to figure out what prefix it is
-func getInstallPrefix() string {
-	if runtime.GOOS != "darwin" {
-		return ""
-	}
-
-	executablePath, err := os.Executable()
-	if err != nil {
-		// TODO: figure out what kind of errors might happen, handle it
-		return ""
-	}
-	cellarPath := filepath.Clean(filepath.Join(resolvePath(executablePath), "../../../.."))
-
-	if !strings.HasSuffix(cellarPath, "Cellar") {
-		// looks like it's not installed via homebrew
-		return ""
-	}
-
-	return filepath.Clean(filepath.Join(cellarPath, "../"))
-}
-
-func resolvePath(path string) string {
-	if res, err := filepath.EvalSymlinks(path); err == nil {
-		return res
-	}
-	return path
-}
 
 func generateRootCmd(cfg *config.Config) *ffcli.Command {
 	// init the log formatter for logrus
@@ -340,9 +132,6 @@ func generateRootCmd(cfg *config.Config) *ffcli.Command {
 			return err
 		}
 		logrus.SetLevel(l)
-		if err = loadTargets(&cfg.Agent); err != nil {
-			return fmt.Errorf("configuration is invalid: %w", err)
-		}
 		return startAgent(&cfg.Agent)
 	}
 
@@ -366,7 +155,7 @@ func generateRootCmd(cfg *config.Config) *ffcli.Command {
 			return nil
 		}
 
-		return exec.Cli(&cfg.Exec, args)
+		return exec.Cli(context.Background(), &cfg.Exec, args)
 	}
 
 	connectCmd.Exec = func(ctx context.Context, args []string) error {
@@ -381,7 +170,7 @@ func generateRootCmd(cfg *config.Config) *ffcli.Command {
 			return nil
 		}
 
-		return exec.Cli(&cfg.Exec, args)
+		return exec.Cli(context.Background(), &cfg.Exec, args)
 	}
 
 	dbmanagerCmd.Exec = func(ctx context.Context, args []string) error {
@@ -425,55 +214,15 @@ func Start(cfg *config.Config) error {
 	return generateRootCmd(cfg).ParseAndRun(context.Background(), os.Args[1:])
 }
 
-func startServer(cfg *config.Server) error {
-	// new a storage with configuration
-	s, err := storage.New(cfg)
-	if err != nil {
-		return fmt.Errorf("new storage: %v", err)
+func resolvePath(path string) string {
+	if res, err := filepath.EvalSymlinks(path); err == nil {
+		return res
 	}
-	atexit.Register(func() { s.Close() })
-
-	// new a direct upstream
-	u := direct.New(s)
-
-	// uploading the server profile self
-	if err := agent.SelfProfile(uint32(cfg.SampleRate), u, "pyroscope.server", logrus.StandardLogger()); err != nil {
-		return fmt.Errorf("start self profile: %v", err)
-	}
-
-	// debuging the RAM and disk usages
-	go printRAMUsage()
-	go printDiskUsage(cfg)
-
-	// new server
-	c, err := server.New(cfg, s)
-	if err != nil {
-		return fmt.Errorf("new server: %v", err)
-	}
-	atexit.Register(func() { c.Stop() })
-
-	// start the analytics
-	if !cfg.AnalyticsOptOut {
-		analyticsService := analytics.NewService(cfg, s, c)
-		go analyticsService.Start()
-		atexit.Register(func() { analyticsService.Stop() })
-	}
-	// if you ever change this line, make sure to update this homebrew test:
-	//   https://github.com/pyroscope-io/homebrew-brew/blob/main/Formula/pyroscope.rb#L94
-	logrus.Info("starting HTTP server")
-
-	// start the server
-	return c.Start()
-}
-
-func startAgent(cfg *config.Agent) error {
-	fmt.Printf("%+v\n", cfg)
-	// TODO: implement agent.
-	return nil
+	return path
 }
 
 func loadTargets(c *config.Agent) error {
-	b, err := os.ReadFile(c.Config)
+	b, err := ioutil.ReadFile(c.Config)
 	if err != nil {
 		return err
 	}
