+++ pkg/cli/cli_linux.go
@@ -1,0 +1,7 @@
+// +build !windows
+
+package cli
+
+func defaultAgentConfigPath() string { return "/etc/pyroscope/agent.yml" }
+
+func getInstallPrefix() string { return "" }
