+++ pkg/exec/exec_test.go
@@ -4,6 +4,8 @@
 package exec
 
 import (
+	"context"
+
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 	"github.com/pyroscope-io/pyroscope/pkg/config"
@@ -18,14 +20,14 @@ var _ = Describe("Cli", func() {
 		Describe("Cli", func() {
 			Context("no arguments", func() {
 				It("returns error", func() {
-					err := Cli(&(*cfg).Exec, []string{})
+					err := Cli(context.Background(), &(*cfg).Exec, []string{})
 					Expect(err).To(MatchError("no arguments passed"))
 				})
 			})
 			Context("simple case", func() {
 				It("returns nil", func() {
 					(*cfg).Exec.SpyName = "debugspy"
-					err := Cli(&(*cfg).Exec, []string{"ls"})
+					err := Cli(context.Background(), &(*cfg).Exec, []string{"ls"})
 					Expect(err).ToNot(HaveOccurred())
 				})
 			})
